chore(test-util): move the steps into a separate package
* move all steps into one package to reuse it and to have a better overview
