+++ test-util/src/main/java/io/zeebe/test/util/bpmn/random/blocks/ExclusiveGatewayBlockBuilder.java
@@ -9,12 +9,14 @@ package io.zeebe.test.util.bpmn.random.blocks;
 
 import io.zeebe.model.bpmn.builder.AbstractFlowNodeBuilder;
 import io.zeebe.model.bpmn.builder.ExclusiveGatewayBuilder;
-import io.zeebe.test.util.bpmn.random.AbstractExecutionStep;
 import io.zeebe.test.util.bpmn.random.BlockBuilder;
 import io.zeebe.test.util.bpmn.random.BlockBuilderFactory;
 import io.zeebe.test.util.bpmn.random.ConstructionContext;
 import io.zeebe.test.util.bpmn.random.ExecutionPathSegment;
 import io.zeebe.test.util.bpmn.random.IDGenerator;
+import io.zeebe.test.util.bpmn.random.steps.StepExpressionIncidentCase;
+import io.zeebe.test.util.bpmn.random.steps.StepPickConditionCase;
+import io.zeebe.test.util.bpmn.random.steps.StepPickDefaultCase;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
@@ -113,155 +115,6 @@ public class ExclusiveGatewayBlockBuilder implements BlockBuilder {
     result.append(blockBuilder.findRandomExecutionPath(random));
 
     return result;
-  }
-
-  // this class could also be called "Set variables when starting the process so that the engine
-  // will select a certain condition"
-  public static final class StepPickConditionCase extends AbstractExecutionStep {
-
-    private final String forkingGatewayId;
-    private final String edgeId;
-
-    public StepPickConditionCase(
-        final String forkingGatewayId, final String gatewayConditionVariable, final String edgeId) {
-      this.forkingGatewayId = forkingGatewayId;
-      this.edgeId = edgeId;
-      variables.put(gatewayConditionVariable, edgeId);
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-        return false;
-      }
-
-      final StepPickConditionCase that = (StepPickConditionCase) o;
-
-      if (forkingGatewayId != null
-          ? !forkingGatewayId.equals(that.forkingGatewayId)
-          : that.forkingGatewayId != null) {
-        return false;
-      }
-      if (edgeId != null ? !edgeId.equals(that.edgeId) : that.edgeId != null) {
-        return false;
-      }
-      return variables.equals(that.variables);
-    }
-
-    @Override
-    public int hashCode() {
-      int result = forkingGatewayId != null ? forkingGatewayId.hashCode() : 0;
-      result = 31 * result + (edgeId != null ? edgeId.hashCode() : 0);
-      result = 31 * result + variables.hashCode();
-      return result;
-    }
-
-    public void removeVariable(final String variable) {
-      variables.remove(variable);
-    }
-  }
-
-  // This class removes the variable set by the `StepPickConditionCase` to make sure an incident is
-  // raised. This same variable can later be provided (through variable update) and the incident can
-  // then be resolved
-  public static final class StepExpressionIncidentCase extends AbstractExecutionStep {
-
-    private final String forkingGatewayId;
-    private final String edgeId;
-    private final String gatewayConditionVariable;
-
-    public StepExpressionIncidentCase(
-        final String forkingGatewayId,
-        final String gatewayConditionVariable,
-        final String edgeId,
-        final StepPickConditionCase pickConditionCase) {
-      this.forkingGatewayId = forkingGatewayId;
-      this.edgeId = edgeId;
-      this.gatewayConditionVariable = gatewayConditionVariable;
-      pickConditionCase.removeVariable(gatewayConditionVariable);
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-        return false;
-      }
-
-      final StepExpressionIncidentCase that = (StepExpressionIncidentCase) o;
-
-      if (forkingGatewayId != null
-          ? !forkingGatewayId.equals(that.forkingGatewayId)
-          : that.forkingGatewayId != null) {
-        return false;
-      }
-      if (edgeId != null ? !edgeId.equals(that.edgeId) : that.edgeId != null) {
-        return false;
-      }
-      return variables.equals(that.variables);
-    }
-
-    @Override
-    public int hashCode() {
-      int result = forkingGatewayId != null ? forkingGatewayId.hashCode() : 0;
-      result = 31 * result + (edgeId != null ? edgeId.hashCode() : 0);
-      result = 31 * result + variables.hashCode();
-      return result;
-    }
-
-    public String getGatewayElementId() {
-      return forkingGatewayId;
-    }
-
-    public String getGatewayConditionVariable() {
-      return gatewayConditionVariable;
-    }
-
-    public String getEdgeId() {
-      return edgeId;
-    }
-  }
-
-  public static final class StepPickDefaultCase extends AbstractExecutionStep {
-
-    private final String forkingGatewayId;
-
-    public StepPickDefaultCase(
-        final String forkingGatewayId, final String gatewayConditionVariable) {
-      this.forkingGatewayId = forkingGatewayId;
-      variables.put(gatewayConditionVariable, "default-case");
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-        return false;
-      }
-
-      final StepPickDefaultCase that = (StepPickDefaultCase) o;
-
-      if (forkingGatewayId != null
-          ? !forkingGatewayId.equals(that.forkingGatewayId)
-          : that.forkingGatewayId != null) {
-        return false;
-      }
-      return variables.equals(that.variables);
-    }
-
-    @Override
-    public int hashCode() {
-      int result = forkingGatewayId != null ? forkingGatewayId.hashCode() : 0;
-      result = 31 * result + variables.hashCode();
-      return result;
-    }
   }
 
   static class Factory implements BlockBuilderFactory {
