+++ test-util/src/main/java/io/zeebe/test/util/bpmn/random/blocks/MessageStartEventBuilder.java
@@ -8,10 +8,10 @@
 package io.zeebe.test.util.bpmn.random.blocks;
 
 import io.zeebe.model.bpmn.builder.ProcessBuilder;
-import io.zeebe.test.util.bpmn.random.AbstractExecutionStep;
 import io.zeebe.test.util.bpmn.random.ConstructionContext;
 import io.zeebe.test.util.bpmn.random.ExecutionPathSegment;
 import io.zeebe.test.util.bpmn.random.StartEventBlockBuilder;
+import io.zeebe.test.util.bpmn.random.steps.StepPublishStartMessage;
 import java.util.Map;
 
 public final class MessageStartEventBuilder implements StartEventBlockBuilder {
@@ -39,43 +39,5 @@ public final class MessageStartEventBuilder implements StartEventBlockBuilder {
     final var pathSegment = new ExecutionPathSegment();
     pathSegment.append(new StepPublishStartMessage(messageName, variables));
     return pathSegment;
-  }
-
-  public static final class StepPublishStartMessage extends AbstractExecutionStep {
-
-    private final String messageName;
-
-    public StepPublishStartMessage(final String messageName, final Map<String, Object> variables) {
-      this.messageName = messageName;
-      this.variables.putAll(variables);
-    }
-
-    public String getMessageName() {
-      return messageName;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-        return false;
-      }
-
-      final StepPublishStartMessage that = (StepPublishStartMessage) o;
-
-      if (messageName != null ? !messageName.equals(that.messageName) : that.messageName != null) {
-        return false;
-      }
-      return variables.equals(that.variables);
-    }
-
-    @Override
-    public int hashCode() {
-      int result = messageName != null ? messageName.hashCode() : 0;
-      result = 31 * result + variables.hashCode();
-      return result;
-    }
   }
 }
