+++ test-util/src/main/java/io/zeebe/test/util/bpmn/random/blocks/NoneStartEventBuilder.java
@@ -8,10 +8,10 @@
 package io.zeebe.test.util.bpmn.random.blocks;
 
 import io.zeebe.model.bpmn.builder.ProcessBuilder;
-import io.zeebe.test.util.bpmn.random.AbstractExecutionStep;
 import io.zeebe.test.util.bpmn.random.ConstructionContext;
 import io.zeebe.test.util.bpmn.random.ExecutionPathSegment;
 import io.zeebe.test.util.bpmn.random.StartEventBlockBuilder;
+import io.zeebe.test.util.bpmn.random.steps.StepStartProcessInstance;
 import java.util.Map;
 
 public final class NoneStartEventBuilder implements StartEventBlockBuilder {
@@ -35,43 +35,5 @@ public final class NoneStartEventBuilder implements StartEventBlockBuilder {
     final var pathSegment = new ExecutionPathSegment();
     pathSegment.append(new StepStartProcessInstance(processId, variables));
     return pathSegment;
-  }
-
-  public static final class StepStartProcessInstance extends AbstractExecutionStep {
-
-    private final String processId;
-
-    public StepStartProcessInstance(final String processId, final Map<String, Object> variables) {
-      this.processId = processId;
-      this.variables.putAll(variables);
-    }
-
-    public String getProcessId() {
-      return processId;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-        return false;
-      }
-
-      final StepStartProcessInstance that = (StepStartProcessInstance) o;
-
-      if (processId != null ? !processId.equals(that.processId) : that.processId != null) {
-        return false;
-      }
-      return variables.equals(that.variables);
-    }
-
-    @Override
-    public int hashCode() {
-      int result = processId != null ? processId.hashCode() : 0;
-      result = 31 * result + variables.hashCode();
-      return result;
-    }
   }
 }
