+++ test-util/src/main/java/io/zeebe/test/util/bpmn/random/steps/AbstractExecutionStep.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.zeebe.test.util.bpmn.random;
+package io.zeebe.test.util.bpmn.random.steps;
 
 import java.util.Collections;
 import java.util.HashMap;
